+++ app/src/main/java/com/eveningoutpost/dexdrip/insulin/Insulin.java
@@ -85,7 +85,7 @@ public abstract class Insulin {
     public ArrayList<Double> getIOBList(int timesliceSize)
     {
         ArrayList<Double> ret = new ArrayList<>();
-        double time = 0;
+        long time = 0;
         double iob = 1;
         while (iob > 1.0/1000000)
         {
